Add first-pass 'promote' functionality
Previously the "promote" subcommand was called "patch".  I thought this could
lead to some confusion.  Typically the source file in a "patch" operation
remains untouched by the patching process (and can be reverted and re-done if
necessary.  However, that's not what happens here.  The idea is to "promote"
(move) settings from the local file to a default file.  And in this case, we
don't want the local file to retain the same settings as the default file (as
that could mask further changes made via version control; the assumption here
is that 'default' is version controlled.)  Therefore, the rename "promote"
seemed more appropriate.

- Implemented first pass of "promote" functionality.  Right now only the
  automatic (non-interactive) version has been implemented.  Started taking
  notes on what interactive mode should look like (borrowing heavily from the
  git "--patch" mode.)
- Allow write_conf() to accept a filename or a stream, like parse_conf()
- Added global EXIT_CODE_* variables for consistent CLI scripting