First bare-bones implementation of 'promote --interactive'
Completed a very rough and ugly interactive version of the promote subcommand.
It only accepts (y/n) at the moment, all changes are listed sequentially with
no means of jumping/skipping/searching/editing.  Diff formatting is unreadable.

Revamped the core diff function `compare_cfgs()` to return namedtuples instead
of plain tuples for the `location` field.  Also made it possible to prevent the
entire file from being returned as a single giant diff by passing in the
`allow_level0` argument.  (Also renumbered the "layers", but this only impacted
the comments.  Now the layer numbers are listed consistently everywhere.)

Added directory support to 'promote'.  If target is a directory, the basename
of the source is simply added to the target.  This allows a user so simply
point to "default" as the target.  (Saves keystrokes when running multiple
promote commands back-to-back.)

Added support for automatic removal of the source file if no content remains
after the promotion.  Also added a CLI option to op-out.

Updated various help command line text as well as comments.