Allow blacklisting of .conf files using a magic string in the comments
Prevent 'ksconf sort -i' from sorting a file by adding 'KSCONF-NO-SORT' at the
top of the file.  (Presently, the string must appear with the first 4k of the
conf file.)  This is helpful for manually edited files with special comment or
special ordering that you do not wish this command to sort.

Disable the message caused by failing to load the optional "argcomplete"
module.  (This was only for debugging.)