Add support for the 'unarchive' command
Created code for the 'unarchive' command.  It supports extracting from both
compressed tarballs (.tar.gz/.tgz/.spl) as well as extracting zip files.
Both install and upgrade operations are supported and automatically detected.
When the destination is a git working tree then a reasonable attempt is made
to make sure the destination is clean to prevent loss of local work.
Several sanity checking levels exist to vary how strict the clean check is.
Also, there are several git modes where the user can control is the app is
simply extracted to the working tree, if the working tree changes should be
staged, and if the staged changes should be automatically committed.

Additionally, the unarchive command supports:  (1) Renaming of the app being
installed.  (A few automatic renaming rules are in place, but primarily this is
a CLI feature.)  (2) Exclusion rules for unwanted content (3) Support for app
that ship with "local" files, (4) renaming of the 'default' folder, this if for
use with the 'ksconf combine' mode.

More work on the 'unarchive' subcommand is likely, but this seems good enough
for a first pass.

Added new git_*() helper functions.

Updated the match_bwlist() function to support Splunk props-style globs: "?"
means any character, "*" means any value within a single directory, and "..."
means any value across directory boundaries.

Added some example argcomplete completer values.  (May need refactored later.)