Expanded CLI help and bash completion
Added completers to all existing subcommands arguments, as appropriate.  Made
"no-op" (fake) completers for when the agrcomplete module is missing, since no
external module dependency is an important design goal of ksconf.

Expand the "description" docs of various subcommands (shown on the help of
'ksconf subcmd --help') and switched over to a pre-formatted style.  Also
attempted to expand the descriptions and reduce the 'help' (which is shown from
the main help page (ksconf --help) so that the main output is more concise.

Create my own argparse formatter (mostly copied for now, may need tweaks)

Greatly expanded the 'combine' docs, since it's a core feature that's a
non-obvious solution.