Initial setup of pre-commit-config.yaml
Setup pre-commit configuration for the local repo.  (This repo can now be used
as a hook provider, and now it uses pre-commit to handle its own hooks.)

Ensure that README.md is always up-to-date by registering make_cli_docs.py as a
pre-commit hook and updated the script to determine if the output file was
updated or not and return an exit code accordingly.  The README.md file should
always be in-sync with the code and any non-staged edits should no leak.

Setup .gitlint file for some basic commit message validation.  (Currently
disabled launching this via pre-commit.  Needs more testing.)