Add conf parsing profiles, a container class, and argparse integration
Updated the parse_conf() interface to accept a 'profile' instead of passing
around a dozen individual parser settings all the time.  For the moment, this
is still implemented as a simple Python dictionary, but eventually this will be
swapped out when the parser itself is overhauled.

Introduced a new 'ConfFileType' class that works as an argparse argument 'type'
that allows an easier integration with with the cli implementation code.  For
example, it's now possible to hand-off simple conf file validation steps to
this new class and therefore skip duplicate implementations of the traditional
"can't read file", and "can't parse file" checks.  Now when either of these
conditions occur, the user is prompted with an error message back from argparse
(before custom CLI code is even called) and consistently across subcommands.

The ConfFileType class works by passing ConfFileProxy objects to the CLI
implementation functions.  These classes also store the appropriate parsing
profile because it's often desirable to vary the parsing features depending on
the nature of the subcommand.  This proxy class provides a simple interface for
reading and writing Splunk .conf files.  Also, because some commands, such as
'minimize' allow a directory to be indicated, instead of a single conf file,
the ConfDirProxy class was created.

Return codes for several subcommands have been tweaked.  This still needs more
work but should at least be a step in the right direction.

The '--duplicate-stanza' and '--duplicate-key' CLI argument have been disabled
(and will most likely be remove and replaced with profile selection options.)