Add '--explode-default' option to minimize
Add feature to allow minimization across not only stanzas of the same name, but
also from the global (or [default] stanza as well.  More testing is needed, but
this is off by default (and probably always will be.)

This can be very helpful when attempting to remove a bunch of default values
after rebuilding a stanza from the output of a REST call or from 'btool list'
where ALL the default values are include.

Also found some weirdness in Splunk's system/default files (such as props.conf)
and added a new unit test so we can handle these cases better.