Support conf files with UTF-8 BOM
Add support for handling UTF-8 byte-order markers.  (Where "support" means
blindly dropping.)  Eventually we'll need proper Unicode (UTF-8) support but
for now we just tread everything as bytes and continue along our merry way in
the trouble-free Python 2 world.  (To be clear, UTF-8 with BOM is a different
encoding type, so this will be a pain, even with proper Unicode support.)