Add automatic build version numbering based on git tags
Git tag information is captured from the 'git describe' command and fed into a
an unversioned _version.py file.  The build version is available as internally
as 'ksconf.version' or from the CLI using 'ksconf --version'.

Additional metadata can/should be added, but this is a big enough win for now.