Expand install docs into INSTALL.md
Added first draft of INSTALL documentation that goes beyond the default
'pip install .' option.  Some installations don't have pip, the right version
of python or even Internet access, and while we can't deal with all the issues,
hopefully we can cover some of the basics and give helpful links.

The biggest addition in the install instructions for using embedded python that
ships with Splunk Enterprise.  There's a number of issues with this, as noted
in the docs, but sometimes it's a good option.

The install docs currently make no mention of how to upgrade.  Fairly straight
forward for the pip use case, but others require more thought.  As of now, NONE
of these options are documented.

I also didn't document the standard 'python setup.py install' option as that
still requires setuptools, which if pip isn't installed (because the Linux
python-devel package is missing, for example), is unlikely to be present.