Create entry_point handling fallback mechanism
Add a new fallback mechanism which required changing the location of the
'ksconf_cmd' entrypoints once again.  The entrypoints are now stored in a
temporary data structure from that we generate (1) the setup() parameters for
packaging, and (2) the fallback mechanism used to load entry points when
external modules are not installed or before the kintyre-splunk-conf package
has been installed for the first time.  (The run_test.py and make_cli_docs.py
scripts are examples of things that weren't working correctly if the package
wasn't installed.)

Additionally, there was a bug with Python 3.4/3.5 on Windows where it ignored
the entry_points defined in 'setup.cfg'.  But it works fine if they are defined
in setup.py.

I also refactored the get_endpoints() code into separate functions per method.