Add 'snapshot' command
Introducing a new prototype command for taking a snapshot of one or more .conf
files and exporting them in an eventually stable export format using JSON. This
can be used for various purposes including capturing a snapshot for purpose of
archival or change control review.  Potentially this format could be used to
ship around conf files when moving full apps or tarballs is undesirable for
some reason (currently no mechanism exists for reconstructing .conf files from
this JSON format, but doing so would be trivial.)

Consider this a functional prototype until further notice.  Feedback welcome,
additional features are planned but not documented anywhere.