Add maturity levels to each subcommand
Add new 'maturity' field to each subcommand to make it easier for users to know
which features are more stable and which ones are still under development.
This allows preview features to be published along side stable ones with a more
clear indication of development process and robustness.  This information is
available by running "ksconf --version"