ConfFileProxy is_file() logic tweak
A fix to ConfFileProxy that doesn't actually fix anything ;-) I think the code
is cleaner, it just wasn't technically necessary for the "is_file()" detection
since ConfFileType is passing in the value explicitly.  Putting this in it's
own commit in case I change my mind.  (See f0b8738f4ff876 for more info)