Add new TermColor helper
Created a new context-managing class as a replacement for the simple
tty_color() function.  I ran into an issue in the field where and unhandled
exception in the diff library (surely caused by something external) left the
terminal in a hard-to-read color.   And that should just never happen.  This
context manager should keep this from being an issue.

Written at 2AM after a 6 hour client working session.  So, do more testing.