Properly support Python ver startup abort
Refactor ksconf.cli and ksconf.__main__ so that ksconf can explain that
you're Python is old and bad in a graceful way that doesn't require me
strip back out all the cool Python 3 stuff we've done so far.

This means that ksconf.__main__ modules should remain Python 2.7
(possibly 2.6, if module.__main__ even worked in that version??) *just*
so we can show a graceful error message if/when somebody tries to run
this on an old version of Python.

TL;DR: Trying to have our cake, and eat it too.